refactor: refactor feedback generation using Factory pattern
