+++ src/main/java/com/nulabinc/zxcvbn/Feedback.java
@@ -1,9 +1,7 @@
 package com.nulabinc.zxcvbn;
 
-import com.nulabinc.zxcvbn.guesses.DictionaryGuess;
 import com.nulabinc.zxcvbn.matchers.Match;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -16,9 +14,6 @@ public class Feedback {
 
   private static final ResourceBundle.Control CONTROL =
       ResourceBundle.Control.getNoFallbackControl(ResourceBundle.Control.FORMAT_DEFAULT);
-
-  private static final List<String> NAME_DICTIONARIES =
-      Arrays.asList("surnames", "male_names", "female_names");
 
   public static final String DEFAULT_SUGGESTIONS_USE_FEW_WORDS =
       "feedback.default.suggestions.useFewWords";
@@ -71,7 +66,7 @@ public class Feedback {
   private final String warning;
   private final String[] suggestions;
 
-  private Feedback(String warning, String... suggestions) {
+  Feedback(String warning, String... suggestions) {
     this.warning = warning;
     this.suggestions = suggestions;
   }
@@ -130,11 +125,11 @@ public class Feedback {
 
   static Feedback getFeedback(int score, List<Match> sequence) {
     if (sequence.size() == 0) {
-      return getFeedbackWithoutWarnings(
+      return FeedbackFactory.getFeedbackWithoutWarnings(
           DEFAULT_SUGGESTIONS_USE_FEW_WORDS, DEFAULT_SUGGESTIONS_NO_NEED_SYMBOLS);
     }
     if (score > 2) {
-      return getEmptyFeedback();
+      return FeedbackFactory.getEmptyFeedback();
     }
     Match longestMatch = sequence.get(0);
     if (sequence.size() > 1) {
@@ -145,142 +140,7 @@ public class Feedback {
       }
     }
 
-    return createMatchFeedback(longestMatch, sequence.size() == 1);
-  }
-
-  private static Feedback getFeedbackWithoutWarnings(String... suggestions) {
-    return new Feedback(null, suggestions);
-  }
-
-  private static Feedback getEmptyFeedback() {
-    return new Feedback(null);
-  }
-
-  private static Feedback createMatchFeedback(Match match, boolean isSoleMatch) {
-    switch (match.pattern) {
-      case Dictionary:
-        return createDictionaryMatchFeedback(match, isSoleMatch);
-      case Spatial:
-        return createSpatialMatchFeedback(match);
-      case Repeat:
-        return createRepeatMatchFeedback(match);
-      case Sequence:
-        return createSequenceMatchFeedback();
-      case Regex:
-        return createRegexMatchFeedback(match);
-      case Date:
-        return createDateMatchFeedback();
-      default:
-        return getFeedbackWithoutWarnings(EXTRA_SUGGESTIONS_ADD_ANOTHER_WORD);
-    }
-  }
-
-  private static Feedback createSpatialMatchFeedback(Match match) {
-    String warning =
-        match.turns == 1
-            ? SPATIAL_WARNING_STRAIGHT_ROWS_OF_KEYS
-            : SPATIAL_WARNING_SHORT_KEYBOARD_PATTERNS;
-    return new Feedback(
-        warning,
-        EXTRA_SUGGESTIONS_ADD_ANOTHER_WORD,
-        SPATIAL_SUGGESTIONS_USE_LONGER_KEYBOARD_PATTERN);
-  }
-
-  private static Feedback createRepeatMatchFeedback(Match match) {
-    String warning =
-        match.baseToken.length() == 1 ? REPEAT_WARNING_LIKE_AAA : REPEAT_WARNING_LIKE_ABCABCABC;
-    return new Feedback(
-        warning, EXTRA_SUGGESTIONS_ADD_ANOTHER_WORD, REPEAT_SUGGESTIONS_AVOID_REPEATED_WORDS);
-  }
-
-  private static Feedback createSequenceMatchFeedback() {
-    return new Feedback(
-        SEQUENCE_WARNING_LIKE_ABCOR6543,
-        EXTRA_SUGGESTIONS_ADD_ANOTHER_WORD,
-        SEQUENCE_SUGGESTIONS_AVOID_SEQUENCES);
-  }
-
-  private static Feedback createRegexMatchFeedback(Match match) {
-    String warning = "recent_year".equals(match.regexName) ? REGEX_WARNING_RECENT_YEARS : null;
-    return new Feedback(
-        warning, EXTRA_SUGGESTIONS_ADD_ANOTHER_WORD, REGEX_SUGGESTIONS_AVOID_RECENT_YEARS);
-  }
-
-  private static Feedback createDateMatchFeedback() {
-    return new Feedback(
-        DATE_WARNING_DATES, EXTRA_SUGGESTIONS_ADD_ANOTHER_WORD, DATE_SUGGESTIONS_AVOID_DATES);
-  }
-
-  private static Feedback createDictionaryMatchFeedback(Match match, boolean isSoleMatch) {
-    String warning = getWarningBasedOnMatch(match, isSoleMatch);
-    List<String> suggestions = generateSuggestions(match);
-    return new Feedback(warning, suggestions.toArray(new String[0]));
-  }
-
-  private static String getWarningBasedOnMatch(Match match, boolean isSoleMatch) {
-    if ("passwords".equals(match.dictionaryName)) {
-      return getPasswordWarning(match, isSoleMatch);
-    }
-
-    if ("english_wikipedia".equals(match.dictionaryName) && isSoleMatch) {
-      return DICTIONARY_WARNING_ENGLISH_WIKIPEDIA_ITSELF;
-    }
-
-    if (NAME_DICTIONARIES.contains(match.dictionaryName)) {
-      return getNameDictionaryWarning(isSoleMatch);
-    }
-
-    return null;
-  }
-
-  private static String getPasswordWarning(Match match, boolean isSoleMatch) {
-    if (isSoleMatch && !match.l33t && !match.reversed) {
-      if (match.rank <= 10) {
-        return DICTIONARY_WARNING_PASSWORDS_TOP10;
-      }
-      if (match.rank <= 100) {
-        return DICTIONARY_WARNING_PASSWORDS_TOP100;
-      }
-      return DICTIONARY_WARNING_PASSWORDS_VERY_COMMON;
-    }
-    if (match.guessesLog10 <= 4) {
-      return DICTIONARY_WARNING_PASSWORDS_SIMILAR;
-    }
-    return null;
-  }
-
-  private static String getNameDictionaryWarning(boolean isSoleMatch) {
-    if (isSoleMatch) {
-      return DICTIONARY_WARNING_ETC_NAMES_THEMSELVES;
-    }
-    return DICTIONARY_WARNING_ETC_NAMES_COMMON;
-  }
-
-  private static List<String> generateSuggestions(Match match) {
-    List<String> suggestions = new ArrayList<>();
-    suggestions.add(EXTRA_SUGGESTIONS_ADD_ANOTHER_WORD);
-
-    CharSequence word = match.token;
-    WipeableString lower = WipeableString.lowerCase(word);
-
-    if (DictionaryGuess.START_UPPER.matcher(word).find()) {
-      suggestions.add(DICTIONARY_SUGGESTIONS_CAPITALIZATION);
-    }
-
-    if (DictionaryGuess.ALL_UPPER.matcher(word).find() && !lower.equals(word)) {
-      suggestions.add(DICTIONARY_SUGGESTIONS_ALL_UPPERCASE);
-    }
-
-    if (match.reversed && match.tokenLength() >= 4) {
-      suggestions.add(DICTIONARY_SUGGESTIONS_REVERSED);
-    }
-
-    if (match.l33t) {
-      suggestions.add(DICTIONARY_SUGGESTIONS_L33T);
-    }
-
-    lower.wipe();
-    return suggestions;
+    return FeedbackFactory.createMatchFeedback(longestMatch, sequence.size() == 1);
   }
 
   private static class ResourceBundleFeedback extends Feedback {
